requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
boto3==1.35.0
celery==5.4.0
redis==5.0.8
httpx==0.27.0
reportlab==4.2.0
pandas==2.2.0
cryptography==41.0.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pymysql==1.1.0
pyodbc==5.1.0
requests==2.31.0
